Iterating on conditions editing
Updated the room checks form to place the deprecated fields at the bottom.
Updated the help callout for conditions with a quick reference, and starting
layout out the ground work for a proper condition editing widget within our
forms framework. Leaving the completion of that for later as it's a bit
involved.